ch: rename components
